Add cache for Main page, update Observer. Add action restore, deleted for Animes and Doramas.

+++ app/Observers/AnimeObserver.php
@@ -3,6 +3,7 @@
 namespace App\Observers;
 
 use App\Models\Anime;
+use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Storage;
 
 class AnimeObserver
@@ -10,12 +11,12 @@ class AnimeObserver
 
     public function created(Anime $anime): void
     {
-        //
+        Cache::forget('main_animes');
     }
 
     public function updated(Anime $anime): void
     {
-        //
+        Cache::forget('main_animes');
     }
 
     public function updating(Anime $anime): void
@@ -31,19 +32,28 @@ class AnimeObserver
 
     public function deleted(Anime $anime): void
     {
-        $anime->ratings()->delete();
-        Storage::disk('anime_posters')->delete($anime->getOriginal('poster'));
-        Storage::disk('anime_covers')->delete($anime->getOriginal('cover'));
+        Cache::forget('main_animes');
     }
 
     public function restored(Anime $anime): void
     {
-        //
+        Cache::forget('main_animes');
     }
 
     public function forceDeleted(Anime $anime): void
     {
-        Storage::disk('anime_posters')->delete($anime->getOriginal('poster'));
-        Storage::disk('anime_covers')->delete($anime->getOriginal('cover'));
+        $anime->ratings()->delete();
+        $anime->genres()->detach();
+        $anime->studios()->detach();
+
+        if ($anime->getOriginal('poster') !== null) {
+            Storage::disk('anime_posters')->delete($anime->getOriginal('poster'));
+        }
+
+        if ($anime->getOriginal('cover') !== null) {
+            Storage::disk('anime_covers')->delete($anime->getOriginal('cover'));
+        }
+
+        Cache::forget('main_animes');
     }
 }
